+++ src/common/Channel.hpp
@@ -4,6 +4,7 @@
 #include "common/FlagsEnum.hpp"
 #include "messages/LimitedQueue.hpp"
 
+#include <QDate>
 #include <QString>
 #include <QTimer>
 #include <boost/optional.hpp>
@@ -36,6 +37,7 @@ public:
         TwitchWhispers,
         TwitchWatching,
         TwitchMentions,
+        TwitchLive,
         TwitchEnd,
         Irc,
         Misc
@@ -96,6 +98,7 @@ public:
     static std::shared_ptr<Channel> getEmpty();
 
     CompletionModel completionModel;
+    QDate lastDate_;
 
 protected:
     virtual void onConnected();