+++ src/common/ChannelChatters.hpp
@@ -7,6 +7,8 @@
 
 #include "lrucache/lrucache.hpp"
 
+#include <QRgb>
+
 namespace chatterino {
 
 class ChannelChatters
@@ -15,7 +17,7 @@ public:
     ChannelChatters(Channel &channel);
     virtual ~ChannelChatters() = default;  // add vtable
 
-    AccessGuard<const UsernameSet> accessChatters() const;
+    SharedAccessGuard<const UsernameSet> accessChatters() const;
 
     void addRecentChatter(const QString &user);
     void addJoinedUser(const QString &user);