+++ src/controllers/notifications/NotificationController.cpp
@@ -6,6 +6,7 @@
 #include "common/QLogging.hpp"
 #include "controllers/notifications/NotificationModel.hpp"
 #include "providers/twitch/TwitchIrcServer.hpp"
+#include "providers/twitch/TwitchMessageBuilder.hpp"
 #include "providers/twitch/api/Helix.hpp"
 #include "singletons/Toasts.hpp"
 #include "singletons/WindowManager.hpp"
@@ -185,6 +186,9 @@ void NotificationController::getFakeTwitchChannelLiveStatus(
             {
                 getApp()->windows->sendAlert();
             }
+            MessageBuilder builder;
+            TwitchMessageBuilder::liveMessage(channelName, &builder);
+            getApp()->twitch2->liveChannel->addMessage(builder.release());
 
             // Indicate that we have pushed notifications for this stream
             fakeTwitchChannels.push_back(channelName);