+++ src/messages/MessageBuilder.cpp
@@ -29,6 +29,53 @@ MessagePtr makeSystemMessage(const QString &text, const QTime &time)
     return MessageBuilder(systemMessage, text, time).release();
 }
 
+MessagePtr makeAutomodInfoMessage(const AutomodInfoAction &action)
+{
+    auto builder = MessageBuilder();
+
+    builder.emplace<TimestampElement>();
+    builder.message().flags.set(MessageFlag::PubSub);
+
+    builder
+        .emplace<ImageElement>(Image::fromPixmap(getResources().twitch.automod),
+                               MessageElementFlag::BadgeChannelAuthority)
+        ->setTooltip("AutoMod");
+    builder.emplace<TextElement>("AutoMod:", MessageElementFlag::BoldUsername,
+                                 MessageColor(QColor("blue")),
+                                 FontStyle::ChatMediumBold);
+    builder.emplace<TextElement>(
+        "AutoMod:", MessageElementFlag::NonBoldUsername,
+        MessageColor(QColor("blue")));
+    switch (action.type)
+    {
+        case AutomodInfoAction::OnHold: {
+            builder.emplace<TextElement>(("Hey! Your message is being checked "
+                                          "by mods and has not been sent."),
+                                         MessageElementFlag::Text,
+                                         MessageColor::Text);
+        }
+        break;
+        case AutomodInfoAction::Denied: {
+            builder.emplace<TextElement>(("Mods have removed your message."),
+                                         MessageElementFlag::Text,
+                                         MessageColor::Text);
+        }
+        break;
+        case AutomodInfoAction::Approved: {
+            builder.emplace<TextElement>(("Mods have accepted your message."),
+                                         MessageElementFlag::Text,
+                                         MessageColor::Text);
+        }
+        break;
+    }
+
+    builder.message().flags.set(MessageFlag::AutoMod);
+
+    auto message = builder.release();
+
+    return message;
+}
+
 std::pair<MessagePtr, MessagePtr> makeAutomodMessage(
     const AutomodAction &action)
 {
@@ -148,8 +195,7 @@ MessageBuilder::MessageBuilder(TimeoutMessageTag,
 
 MessageBuilder::MessageBuilder(TimeoutMessageTag, const QString &username,
                                const QString &durationInSeconds,
-                               const QString &reason, bool multipleTimes,
-                               const QTime &time)
+                               bool multipleTimes, const QTime &time)
     : MessageBuilder()
 {
     QString fullText;
@@ -178,12 +224,6 @@ MessageBuilder::MessageBuilder(TimeoutMessageTag, const QString &username,
         text.append("has been permanently banned");
     }
 
-    if (reason.length() > 0)
-    {
-        text.append(": \"");
-        text.append(parseTagString(reason));
-        text.append("\"");
-    }
     text.append(".");
 
     if (multipleTimes)
@@ -343,7 +383,7 @@ MessageBuilder::MessageBuilder(const AutomodUserAction &action)
         break;
 
         case AutomodUserAction::RemovePermitted: {
-            text = QString("%1 removed %2 as a permitted term term on AutoMod.")
+            text = QString("%1 removed %2 as a permitted term on AutoMod.")
                        .arg(action.source.name)
                        .arg(action.message);
         }