+++ src/messages/MessageBuilder.hpp
@@ -11,6 +11,7 @@ struct BanAction;
 struct UnbanAction;
 struct AutomodAction;
 struct AutomodUserAction;
+struct AutomodInfoAction;
 struct Message;
 using MessagePtr = std::shared_ptr<const Message>;
 
@@ -25,6 +26,7 @@ MessagePtr makeSystemMessage(const QString &text);
 MessagePtr makeSystemMessage(const QString &text, const QTime &time);
 std::pair<MessagePtr, MessagePtr> makeAutomodMessage(
     const AutomodAction &action);
+MessagePtr makeAutomodInfoMessage(const AutomodInfoAction &action);
 
 struct MessageParseArgs {
     bool disablePingSounds = false;
@@ -44,8 +46,7 @@ public:
     MessageBuilder(TimeoutMessageTag, const QString &systemMessageText,
                    int times, const QTime &time = QTime::currentTime());
     MessageBuilder(TimeoutMessageTag, const QString &username,
-                   const QString &durationInSeconds, const QString &reason,
-                   bool multipleTimes,
+                   const QString &durationInSeconds, bool multipleTimes,
                    const QTime &time = QTime::currentTime());
     MessageBuilder(const BanAction &action, uint32_t count = 1);
     MessageBuilder(const UnbanAction &action);