+++ src/widgets/splits/SplitInput.cpp
@@ -26,6 +26,7 @@
 #include <QPainter>
 
 namespace chatterino {
+const int TWITCH_MESSAGE_LIMIT = 500;
 
 SplitInput::SplitInput(Split *_chatWidget)
     : BaseWidget(_chatWidget)
@@ -619,6 +620,14 @@ void SplitInput::editTextChanged()
     if (text.length() > 0 && getSettings()->showMessageLength)
     {
         labelText = QString::number(text.length());
+        if (text.length() > TWITCH_MESSAGE_LIMIT)
+        {
+            this->ui_.textEditLength->setStyleSheet("color: red");
+        }
+        else
+        {
+            this->ui_.textEditLength->setStyleSheet("");
+        }
     }
     else
     {